Refactor round
