+++ src/computed-properties.js
@@ -1,0 +1,7 @@
+export function createComputedPropertiesMiddleware(references) {
+  return store => next => action => {
+    references.internals.computedState.currentState = store.getState();
+    references.internals.computedState.isInReducer = true;
+    return next(action);
+  };
+}
