+++ src/create-store.js
@@ -12,6 +12,8 @@ import {
   createPersistenceClearer,
   rehydrateStateFromPersistIfNeeded,
 } from './storage';
+import { createComputedPropertiesMiddleware } from './computed-properties';
+import { createListenerMiddleware } from './listeners';
 
 export default function createStore(model, options = {}) {
   const {
@@ -34,9 +36,20 @@ export default function createStore(model, options = {}) {
     };
   };
 
+  const references = {};
+
   let modelDefinition = bindReplaceState(model);
   let mockedActions = [];
-  const references = {};
+
+  const persistKey = targetPath => `[${storeName}]@${targetPath.join('.')}`;
+  const persistor = createPersistor(persistKey, references);
+  const persistMiddleware = createPersistMiddleware(persistor, references);
+  const clearPersistance = createPersistenceClearer(persistKey, references);
+
+  const replaceState = nextState =>
+    references.internals.actionCreatorDict['@action.easyPeasyReplaceState'](
+      nextState,
+    );
 
   const bindStoreInternals = (state = {}) => {
     references.internals = createStoreInternals({
@@ -49,40 +62,6 @@ export default function createStore(model, options = {}) {
     });
   };
 
-  bindStoreInternals(initialState);
-
-  const replaceState = nextState =>
-    references.internals.actionCreatorDict['@action.easyPeasyReplaceState'](
-      nextState,
-    );
-
-  const persistKey = targetPath => `[${storeName}]@${targetPath.join('.')}`;
-  const persistor = createPersistor(persistKey, references);
-  const persistMiddleware = createPersistMiddleware(persistor, references);
-  const clearPersistance = createPersistenceClearer(persistKey, references);
-
-  const listenerActionsMiddleware = () => next => action => {
-    const result = next(action);
-    if (
-      action &&
-      references.internals.listenerActionMap[action.type] &&
-      references.internals.listenerActionMap[action.type].length > 0
-    ) {
-      const sourceAction = references.internals.actionCreatorDict[action.type];
-      references.internals.listenerActionMap[action.type].forEach(
-        actionCreator => {
-          actionCreator({
-            type: sourceAction ? sourceAction.type : action.type,
-            payload: action.payload,
-            error: action.error,
-            result: action.result,
-          });
-        },
-      );
-    }
-    return result;
-  };
-
   const mockActionsMiddleware = () => () => action => {
     if (action != null) {
       mockedActions.push(action);
@@ -90,17 +69,11 @@ export default function createStore(model, options = {}) {
     return undefined;
   };
 
-  const computedPropertiesMiddleware = store => next => action => {
-    references.internals.computedState.currentState = store.getState();
-    references.internals.computedState.isInReducer = true;
-    return next(action);
-  };
-
   const easyPeasyMiddleware = [
-    computedPropertiesMiddleware,
+    createComputedPropertiesMiddleware(references),
     reduxThunk,
     ...middleware,
-    listenerActionsMiddleware,
+    createListenerMiddleware(references),
     persistMiddleware,
   ];
 
@@ -117,6 +90,8 @@ export default function createStore(model, options = {}) {
           name: storeName,
         })
       : reduxCompose);
+
+  bindStoreInternals(initialState);
 
   const store = reduxCreateStore(
     references.internals.reducer,
